Made some changes to serverFilters, cleaned it up to make it more readable. Changed screening list to <a> tags instead of onClick functions. Also made some design changes.
